코드 리펙토링 및 Currency 의존성 문제 발생

+++ src/main/java/com/example/assignment/DTO/AuthorRequestDto.java
@@ -5,5 +5,5 @@ import lombok.Getter;
 @Getter
 public class AuthorRequestDto {
     private String name;
-    private Long birthday;
+    private int birthday;
 }
